Do not plot the last time parameter.
The last time parameter is always affected by various artifacts such as
mismapping, copy-number variantions and balancing selection. PSMC is unable to
estimate that accurately. However, users not familiar with PSMC may make wrong
conclusions based on that. It is harder for me to explain what is happening. To
avoid the complication, I now skip the last parameter when plotting.